[Refactor] component 파일로 분리
